Twiss Elements: Select ALL (#492)
We noticed that in the twiss elements the hkicker and vkicker magnets are missing.
Instead of adding each class individually (and potentially missing some) we could just do a select, flag=twiss, clear; and output all. The file-size for the LHC changes from 7.1MB to 8.3MB which is acceptable, imho.

If we really cared about the file-size we could remove the Drifts (pattern="^[^D].*" probably).

It seems the issue with IP2/8 was due to the model_elements not containing some elements. So this PR fixes that, and tests the IP2/8 data that we tried in the CCC.

---------

Co-authored-by: Joshua Gray <61458311+jgray-19@users.noreply.github.com>
Co-authored-by: jgray-19 <jgray-19@users.noreply.github.